requirements.txt
djangorestframework==3.11.1
Django==3.1